Complete exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -56,6 +56,16 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token
+      || !decodedToken.id
+      || blog.author.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
